refactor: moved client socket to zustand store
