+++ package.json
@@ -22,7 +22,8 @@
       "react": "^19.0.0",
       "react-dom": "^19.0.0",
       "socket.io": "^4.8.1",
-      "socket.io-client": "^4.8.1"
+      "socket.io-client": "^4.8.1",
+      "zustand": "^5.0.3"
    },
    "devDependencies": {
       "@eslint/eslintrc": "^3",
