--- src/app/games/WebSocket.tsx
@@ -1,9 +1,0 @@
-'use client';
-
-import { useWebSocket } from '@/hooks/useWebSocket';
-
-export default function WebSocket() {
-   useWebSocket(true);
-
-   return null;
-}
