+++ src/app/games/layout.tsx
@@ -1,10 +1,3 @@
-import WebSocket from '@/app/games/WebSocket';
-
 export default function Layout({ children }: { children: React.ReactNode }) {
-   return (
-      <>
-         <WebSocket />
-         {children}
-      </>
-   );
+   return children;
 }
